fix: Move WalletsStat component to shared
