+++ web/src/components/shared/sharedLocales/ru.json
@@ -449,5 +449,25 @@
   "userAd.empty.marketLink": "Перейдите в маркет",
   "userAd.empty.title2": "чтобы посмотреть другие объявления или создайте своё объявление",
   "Num": "№",
-  "userAd.adsUpdatedToOften": "Вы редактируете объявление слишком часто. Повторите попытку через несколько секунд"
+  "userAd.adsUpdatedToOften": "Вы редактируете объявление слишком часто. Повторите попытку через несколько секунд",
+
+  "walletsStat.title": "Bitzlato статистика кошельков",
+  "walletsStat.titleShort": "Статистика кошельков",
+  "walletsStat.code": "Код",
+  "walletsStat.withdrawalFee": "Комиссия",
+  "walletsStat.minDeposit": "Мин. ввод",
+  "walletsStat.minWithdrawal": "Мин. вывод",
+  "walletsStat.withdrawEnabled": "Вывод включен",
+  "walletsStat.depositEnabled": "Ввод включен",
+  "walletsStat.withdrawalQueue": "Очередь выводов",
+  "walletsStat.withdrawalErrors": "Ошибки выводов",
+  "Table headers": "Заголовки таблицы",
+  "walletsStat.notice.code": "Код криптовалюты",
+  "walletsStat.notice.withdrawalFee": "Комиссия за вывод в криптовалюте",
+  "walletsStat.notice.minDeposit": "Минимальная сумма ввода в криптовалюте",
+  "walletsStat.notice.minWithdrawal": "Минимальный вывод в криптовалюте",
+  "walletsStat.notice.withdrawEnabled": "Вывод административно включен",
+  "walletsStat.notice.depositEnabled": "Ввод административно включен",
+  "walletsStat.notice.withdrawalQueue": "Количество выводов в очереди",
+  "walletsStat.notice.withdrawalErrors": "Количество неуспешных выводов за 24 часа"
 }
