+++ web/src/translations/en.ts
@@ -1677,24 +1677,4 @@ export const en = {
   Save: 'Save',
 
   'Start P2P Trading': 'Start P2P Trading',
-
-  'walletsStat.title': 'Bitzlato wallets statistics',
-  'walletsStat.titleShort': 'Wallets statistics',
-  'walletsStat.code': 'Code',
-  'walletsStat.withdrawalFee': 'Withdrawal fee',
-  'walletsStat.minDeposit': 'Min deposit',
-  'walletsStat.minWithdrawal': 'Min withdrawal',
-  'walletsStat.withdrawEnabled': 'Withdraw enabled',
-  'walletsStat.depositEnabled': 'Deposit enabled',
-  'walletsStat.withdrawalQueue': 'Withdrawal queue',
-  'walletsStat.withdrawalErrors': 'Withdrawal errors',
-  'Table headers': 'Table headers',
-  'walletsStat.notice.code': 'Cryptocurrency code',
-  'walletsStat.notice.withdrawalFee': 'Withdrawal fee',
-  'walletsStat.notice.minDeposit': 'Minimum deposit amount',
-  'walletsStat.notice.minWithdrawal': 'Min withdraw amount',
-  'walletsStat.notice.withdrawEnabled': 'Withdraw enabled',
-  'walletsStat.notice.depositEnabled': 'Deposit enabled',
-  'walletsStat.notice.withdrawalQueue': 'Withdrawal queue',
-  'walletsStat.notice.withdrawalErrors': 'Withdrawal errors for 24 hours',
 };
